requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.9
pgvector>=0.2.1
celery>=5.3.4
redis>=5.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
prometheus-client>=0.17.1
pydantic>=2.4.2
aioredis>=2.0.1
openai>=1.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
